+++ lib/engine/renderer/renderer.dart
@@ -3,4 +3,5 @@ import 'package:pulsar_web/pulsar.dart';
 abstract class Renderer {
   void mount(PulsarNode root);
   void update(PulsarNode prev, PulsarNode next);
+  void unmount();
 }