+++ lib/engine/router/router_component.dart
@@ -2,17 +2,18 @@ import 'package:pulsar_web/pulsar.dart';
 
 class RouterComponent extends Component {
   static RouterComponent? _instance;
+
   final Router router;
   final Component Function(Component page) layout;
+
+  Component? _currentPage;
 
   RouterComponent({required this.router, required this.layout});
 
   @override
   void onMount() {
     _instance = this;
-    window.onPopState.listen((_) {
-      update();
-    });
+    window.onPopState.listen((_) => update());
   }
 
   static void notifyNavigation() {
@@ -22,9 +23,19 @@ class RouterComponent extends Component {
   @override
   PulsarNode render() {
     final path = window.location.pathname;
-    final component = router.resolve(path);
-    final wrapped = layout(component);
-    return ComponentNode(component: wrapped, key: path);
+
+    final resolved = router.resolve(path);
+
+    // 👇 SOLO cambia si el tipo cambia
+    if (_currentPage == null ||
+        _currentPage.runtimeType != resolved.runtimeType) {
+      _currentPage = resolved;
+    }
+
+    return ComponentNode(
+      component: layout(_currentPage!),
+      // 👇 SIN key
+    );
   }
 }