+++ lib/types.dart
@@ -1,3 +1,5 @@
 import 'package:pulsar_web/pulsar.dart';
 
 typedef ComponentBuilder = Component Function();
+
+typedef ParamsContext = Map<String, String>;