+++ web/main.dart
@@ -1,5 +1,5 @@
-import 'package:pulsar_web/pulsar.dart';
-import 'components/counter/counter_app.dart';
+import 'package:pulsar_web/utils/mount_app.dart';
+import './components/counter/counter_app.dart';
 
 void main() {
   mountApp(CounterApp(), selector: "#app");